python: Added custom driver for multisim
This file provides an example for how to work with the statistic
return values in gem5. Note that the caching mechanism is fairly
rudimentary, I would like to (in the future) validate that the
cached results correspond to the ones you want to run by comparing
the generated .ini files for each simulation.

There should also be a cli arg that allows you to force rerun
all of the simulations, but adding that in a normal way seems to
break everything.