stdlib: RISCVMatchedBoard: Remove sf-pdma from the platform
There is an intention to keep the RISCVMatchedBoard configuration stable,
but if we leave it as it is, the ports of the newly added sf-pdma device
on the HiFive platform will be in an unconnected state, which will cause
the simulation to crash.
The easiest way to fix this is just to remove the DMA device from the
platform for this board.

Signed-off-by: Nikita Proshkin <nic.proshkin@gmail.com>