cpu: Fix Clang compilation error for TAGE-SC-L
PR #499 introduced a shadowing of virtual method
`TAGEBase::updatePathAndGlobalHistory` in `TAGE_SC_L_TAGE` that should
have been an overload. This shadow method works in the context that it
is currently used, but causes a warning-as-error in Clang, which is
more proactive about warning of potentially unwanted shadowing.

This patch matches the prototype to overload the existing virtual
method, avoiding the error in Clang and fixing the build.

Change-Id: If4296f9b6d6154066567a58dc9cbdcd4fc44f425