arch-riscv: Fix atomic ops on big endian hosts
RISC-V atomic operations pass a functor which operates on the data.
Previously, this functor ignored the endianness of the host and just
directly operated on the data. This change ensures that when reading
data from the guest memory into host variables that we set the
endianness correctly.

A small number of these operations were tested on a BE host by ensuring
the execution trace was the same on both the BE host and a LE host.

Signed-off-by: Jason Lowe-Power <jason@lowepower.com>