configs,mem-ruby: Fix Directory_Controller fallback logic
This change corrects the instantiation of the Directory_Controller when
using Ruby protocols. The new behavior attempts to use the
protocol-specific controller class (e.g.,
`MI_example_Directory_Controller`) as required when running with
multiple protocols compiled (typically when using the ALL gem5 binary).
If the protocol-specific controller is not found (the case when working
with a binary compiled to a single protocol), the code falls back to
using the generic Directory_Controller.

This fix resolves an issue that caused the gem5 nightly tests to fail,
as seen in GitHub Actions Run #14650118937.