arch-sparc: Fix SyntaxWarnings in sparc/isa/base.isa
In `re.sub`, the literal `\g<FOO>` is used to reference a capture
group in the regexp by name/number. In ordinary Python strings `\g`
is not a valid escape sequence, and so is passed through directly.
In Python 3.12 this becomes a SyntaxWarning. Make explicit the fact
that we really want the two characters `\g` with no unescaping by
using raw strings.

Change-Id: I2ca8cf3c37e0d6e27baec661ee2c9fb840fb2432