tests: modifications to varying length multisim test
This commit removes the NPB size A and size S workloads from this
test to reduce the amount of time it takes to run the multisim.
tests. It also refactors the code for the RISCV boots and adds code
that should replace the current code when the appropriate resources
have been uploaded.