misc: Add git-clang-format to pre-commit
This commit enables automatic formatting of C/C++ code using
`clang-format` via the `git-clang-format` tool. The `git-clang-format`
tool has the advantage that it only re-formats the changed lines of a
git commit, rather than full files. Therefore it can be used to
incrementally introduce automatic formatting without requiring
re-formatting of the whole repository.

This commit is designed to help gem5 developers by automating part of
the code-review process (adherence to the gem5 style guide).

The `clang-format` tools now become a requirement to work with gem5.
They can be installed on Ubuntu using the command:
    `sudo apt install clang-format`

The `.clang-format` file introduced here is generated by ChatGPT to
closely match the gem5 Style Guide [1]. It is recommended that this
file is replaced by the `.clang-format` proposed in PR #362 [2] by
@trollet91 and other gem5 contributors.

[1] https://www.gem5.org/documentation/general_docs/development/coding_style/
[2] https://github.com/gem5/gem5/pull/362

Change-Id: Ifc1c4a29e954a73b64d64d02124c01ca7f48be47