arch-riscv: Fix timing walk exit path on early fault
Previously, a timing walk could only generate a fault
after sendPackets() had been called at least once.

PR #1387 added a check at the start of walkOneStage()
which can generate a pagefault before sendPackets() is
called.

In that case, the translation was wrongly marked delayed,
which did not allow for its deletion and subsequently
caused a hang.

This commit fixes the above by not marking the translation
delayed in these cases and removing the relevant WalkerStates
when needed.