dev: Export reg bank's offsetMap
This can be useful in some cases. For example, we have a bank of double
buffered registers and on some condition, we need to iterate through all
registers in the bank to sync the buffered state to active state.

Instead of maintaining another list of register references, having this
field exported would help simplify the implementation.