dev: Make DmaPort can be inherited
DmaPort is a base class for modeling a port with DMA ability. It's
possible that the port needing to support different kinds of protocol
with different kinds of metadata. For example for modeling AXI, we need
to add AxCache and AxDomain in the request.

To make the base class be able to extend, I made the private fields to
protected. Secondly, I made DmaReqState::createPacket be able to
override which allows creating a packet with extra information.

Change-Id: I32d4ab1b6388c460d80769969bbbee3b824d2c50