mem: refactor Bridge to allow derived class
This commit splits Bridge into BridgeBase and Bridge. BridgeBase
contains the implementation to handle packets passing through a bridge,
like Bridge before. But instead of relying on a fixed set of address
ranges, it allows other class to derived from it to implement the
functions getAddrRanges() and recvRangeChange() and so the set of
address ranges can be either fixed or dynamic depending on the derived
class implementation.

The functions canPassThrough() and setRejectedData() can also be
implemented by derived classes, allowing to block some ranges and
directly respond to them from the bridge. This is usefull to let a
bridge respond with an error code when no device is responsible for a
given range behind the bridge.