tests: fix kernel panic on kvm switch tests
The resource used for kvm switch tests as kernel arguments set to force
the "root=/dev/sda2". The kvm switch adds some kernel arguments, but
use the X86Board default kernel arguments. The argument is then set to
"root=/dev/hda1" and a kernel panic occurs as it cannot mount root fs.

This commits fix that by getting the kernel arguments from the workload
instead of the default ones.