cpu-o3: Print OpClass name instead of number for debug purposes
A recent PR [1] introduced a panic to catch the case where
the O3 model does not have a capable FU to execute a
particular instruction.

At the moment we exit simulation and we print the OpClass
enum number. This is not particularly useful for a user
as it requires the user to manually extract the OpClass
name from the number [2] and it can be a bit tedious.

With this PR we print the OpClass name directly

[1]: https://github.com/gem5/gem5/pull/1516
[2]: https://github.com/gem5/gem5/blob/stable/src/cpu/FuncUnit.py

Change-Id: I5a9aaa7c569e67a87400d240b40e9d2a9147cad8
Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>