python: Prevent SimObject from sharing reference of class member.
The SimObject clones a member if it is a SimObject or a vector
of SimObjects. However, if the member is an empty list. The
current implementation does not clone this member, causing the
instance sharing the reference of the class member.

It causes the following behavior:

a = Obj1()
a.foo.append(Obj2)

b = Obj1()
len(b.foo) == 1 # True

This commit handles the empty list case. Note that, it invoke
the value's constructor because the member could be a derived
class of a list (e.g. VectorParamValue).