dev: refactor PCI configuration to use RegisterBank
This commit refactors PCI configuration space to use RegisterBank. For
that, the configuration header and capabilities of a device are now child
classes of RegisterBank and stored in a common map with their base
offset as key.

This allows to remove the need of extending readConfig and writeConfig
function in PciDevice subclasses. Actions that are specifics to a
register (e.g. sending range change on BAR writes) can be implemented
using writer/reader from Register class.

Capabilities are also now their own classes, instead of being parameter
in PciDevice which weren't always used.