dev: Fix PCI host bridge with no range from up
PCI host bridge was intended to be connected to ports which returns a
correct set of address ranges when calling getAddrRanges(). But
RubyPort::MemResponsePort does return an empty list, making use of that
function when connected to it, which is the case for system using Ruby
hierarchy.

This commit change the downstream to upstream bridge to be connected to
the default port of PciBus. But, as that port was connected to an
IsaFake to handle configuration access error, that one requires to be
moved. So PciBus now contains a new port for that which will act like a
second default port but only for the configuration ranges.

Also, as down to up bridge is now connected to default, a simple Bridge
class can be used. And so PciOneWayBridge can be renamed to
PciUpDownBridge and simplified as it doesn't require to remove undesired
ranges responded by the down to up bridge.