configs, cpu-o3: Implement a distributed InstructionQueue
This commit is implementing a distributed InstructionQueue.

Prior to this patch it was only possible to have a unified instruction
queue for the O3 CPU: every renamed/dispatched instruction was allocated
in a single queue.

With this patch we define a IQUnit SimObject (which is basically a
single IQ) and we allow the O3CPU to have a vector of IQs to reflect the
presence of multiple instruction queues.

A IQUnit is tighlty coupled with its own FUPool: it can only store
instructions executable by one of the FU in the pool.
This also means we don't have a single fuPool anymore, but a collection
of pools (fuPool is a IQUnit Param)

By default, the unified design is in place. To adopt the multi-IQs
design, one simply has to provide a different vector to the O3CPU

Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>
Change-Id: Ic3923ab779bdb2e957fc5e7a4c5881f9e685aa81