sim,cpu: Make SimObject imports consistent in SimObject files
In SimObject description files, you often need to import other SimObjects
to use as parameter values. Note that these are only used as parameters,
so we don't need the C++ backend at this point.
The problem is that when there is a SimObject description file with multiple
SimObjects declared, it's hard to know what file you should import *from*.
This change begins to make a consistent decision to import from
m5.objects.<SimObjectName> instead of from the file that it was declared in.
This works because all SimObjects end up in m5.objects.<SimObjectName> modules.

Signed-off-by: Jason Lowe-Power <jlowepower@google.com>