misc: multiple stats outputs with --stats-file
This commit allows multiple stats output files to be specified
using --stats-file. This is useful if for exampe we want to
output statistics in multiple different formats.
For example:
"./build/ALL/gem5.opt --stats-file json://stats1.json
--stats-file stats.txt test.py"
Would dump stats in both text and json formats.
If the argument isn't specified we use stats.txt as the
default location for stats dumps.

Change-Id: I049b237ece3e0e89a78bfad19ca071dff8940d5c
Reviewed-by: Giacomo Travaglini <giacomo.travaglini@arm.com>