cpu-o3: Instant ROB squash
The squashWidth limits the number of instructions squashed
within on cycle. In case of a full, large ROB this can incure tens
of cycles during which the rename stage is blocked. E.g. with
160 instructions in the ROB and squash width of 8 rename will be
blocked 20 cycles.
With this change users can set the squash width to a negative number
and then all instructions will be squashed immediatly.
In a real implementation it is more likely the case where this can
simply be done with updating a pointer.