mem-ruby: Update SLICC to generate only shared files
After the "all protocols" change, there are two types of generated files
in SLICC. There are the normal protocol files and the "shared" protocol
files. The shared files are files that are generated *exactly the same*
across all protocols. Currently, we generate these files for each protocol
that is included in the gem5 binary. (The reason we have to is because
SLICC needs to process these files every time.) This change makes it
possible out output *only* these shared files by allowing no protocol
slicc file.

These "shared" files are the only files that should be referenced by
the generic Ruby files. An exception is the CHI protocol which has
some C++ files that directly reference CHI protocol generated files.

Note: this is not currently used in SCons.

Signed-off-by: Jason Lowe-Power <jlowepower@google.com>