mem-ruby: Update Ruby Network to use new-style stats
Ruby is one of the last places in the codebase that are still using
the old way of initializing stats. This change updates the network
and simple network with the new style stats.

This change also fixes a bug where the Switch stats were not reset.

There's one wrinkle. The stats can't be initialized until after ports
are connected and all of the internal/external links have been
created. To get around this, we override the regStats functions
in the stats objects.

*User-facing changes*
- All of the modified stats now have descriptions and units

Signed-off-by: Jason Lowe-Power <jlowepower@google.com>