base-stats: Add lastCommitPC to ThreadStateStats
This commit adds a new statistic, lastCommitPC, to the ThreadStateStats
structure in thread_state.hh. This statistic tracks the program counter
(PC) of the last instruction that was committed by the CPU. We have also
changed simple, O3, and minor CPU models to update this statistic
whenever an instruction is committed.

This addition provides valuable insight into the execution flow of the
simulated CPU, allowing users to monitor the last executed instruction's
address. This can be particularly useful when used with periodicStatDump
to trace sampled execution points over time.

Signed-off-by: Yangyu Chen <cyy@cyyself.name>