arch-x86: implement big movfp micro-op variant
Fix #2778 by implementing a "big" variant for the `movfp`
micro-op, akin to `mov` micro-op's big variant, which
does not depend on the destination FP register's value.
Use the big variant when the data size is 8; otherwise,
fall back to the original implementation which reads
read-modify-writes the destination register.

The IPC for the microbenchmark in #2778 is 2.896672,
up from 1.054388.

Change-Id: I66a68fc42af11e0ab21a6fc6301a748b3335b946