mem-ruby: Add backpressure mechanism in TlmController
This patch implements p-credit handling in the TlmGenerator
to model backpressure from downstream components
(e.g. the HNF).

If a transaction cannot be accepted from the HNF, a RetryAck
will be forwarded back. The requester will have to wait
until a p-credit has been granted to it (via the PCrdGrant
response message).

A TlmGenerator test suite will probably fail at the moment whenever
it receives a RetryAck. For example if a transaction is a ReadUnique,
the test is possibly expecting a CompData as a response.
We could consider handling RetryAck in the python suite file, however
this comes with the following issues:

1) Complication of the suite tests (need to explicitly handle a
transaction retry)

2) The PCrdGrant message is not linked to a specific transaction
(txn_id field is invalid) and as such handling it in a transaction
python object adds extra complications

Change-Id: I2b06dc850c385b41919d1ed57f467d4bb68222d7
Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>
Reviewed-by: Andreas Sandberg <andreas.sandberg@arm.com>